Really implemented a Predicate type
As ctypes does not support 'typedef' defines , it is impossible to extract
the types that are defined using 'typedef', hence why they have to be emulated.

The simple types such as:
"typedef int Precedence;"
are easy - I just create a class with 'c_int' as parent.

However types such as:
"typedef int (*InterceptionPredicate)(InterceptionDevice device);"
are a bit more of a challange.

To emulate it I had to first understand what that syntax meant. Frankly I have not
completed that, yet. But from what I understood, it must be an object that either
points directly to a procedure or a pointer to an object that is callable.

The simplest solution was:
Predicate = CFUNCTYPE( c_int, c_int )

However that is not really a type, so you couldnt do things like strict type checking
( i.e. CFUNCTYPE(c_int,c_int)(function)==Predicate(function) would be evaluated as True)
So obviously I needed a class that:
1. can be passed as an argument to C functions - must have _as_parameter_ attribute
(also must have a from_parameter() method in order to use the class in function prototypes)
2. can be called from python - must have __call__ method

So I made the constructor and i made it work with C functions and any python object that is callable.
And added the 2 trivial methods.

However that was not enough as this way you could not call set_filter
with passing a function that was not a Predicate.

So I wrapped the set_filter function so that if the argument is not already a Predicate,
a Predicate object would be constructed from the argument. However, since the constructed
object is not given to the caller, I didnt want to create new objects everytime the function is called.
So i wrapped the Predicate class, with a cache.