Added a 'StrokeArray' type
'StrokeArray' allows you to make arrays of 'Strokes'.
This(along with other minor tweaks) allows for sending/recieving multiple strokes at once.

Also added a 'StrokeProto' type and made 'Stroke' and 'StrokeArray' inherit it.
Now 'StrokeProto' is used in the descprion of the arguments of the
exposed functions ( the function prototypes ), so both instances of 'Stroke' and 'StrokeArray'
can be passed as arguments.